--- a/src/layout.rs
+++ b/src/layout.rs
@@ -217,7 +217,7 @@ impl<'a> Layout<'a> {
     ///     });
     /// assert_eq!(
     ///     chunks,
-    ///     vec![
+    ///     &[
     ///         Rect {
     ///             x: 2,
     ///             y: 2,
@@ -244,7 +244,7 @@ impl<'a> Layout<'a> {
     ///     });
     /// assert_eq!(
     ///     chunks,
-    ///     vec![
+    ///     &[
     ///         Rect {
     ///             x: 0,
     ///             y: 0,