--- a/src/lib.rs
+++ b/src/lib.rs
@@ -160,6 +160,8 @@
 //! default the computed layout tries to fill the available space completely. So if for any reason
 //! you might need a blank space somewhere, try to pass an additional constraint and don't use the
 //! corresponding area.
+#![feature(test)]
+extern crate test;
 
 pub mod backend;
 pub mod buffer;