## Main changes
I found that the current implementation of `Layout` was very heavy on allocations and clones, especially considering you're expected to use it plus the `split()` function every frame. So, I made some changes to allow `Layout` to work more on the stack.

This is what I changed `Layout` to
```rust
#[derive(Debug, Clone, PartialEq, Eq, Hash)]
pub struct Layout<'a> {
    direction: Direction,
    margin: Margin,
    constraints: &'a [Constraint],
    /// Whether the last chunk of the computed layout should be expanded to fill the available
    /// space.
    expand_to_fill: bool,
}
```
This replaces the `constraints` field with a reference to a slice, instead of a `Vec<Constraint>`. Considering that most implementations using `Layout` use static arrays anyway, this seemed like a no brainer, until I tried to use the `LAYOUT_CAHCE`. It is impossible to store this new `Layout` inside of the `LAYOUT_CACHE` because that would require that you borrow the given `Layout` for `'static`. My solution to the problem was only storing the hash of the `Layout` into the cache as opposed to moving the entire `Layout`.
```rust
macro_rules! hash_layout {
    ($self:expr, $area:expr) => {{
        let mut to_hash = ahash::AHasher::default();
        $area.hash(&mut to_hash);
        $self.margin.hash(&mut to_hash);
        $self.expand_to_fill.hash(&mut to_hash);
        $self.direction.hash(&mut to_hash);
        $self.constraints.iter().copied().for_each(|f| match f {
            Constraint::Max(max) => to_hash.write_u16(max),
            Constraint::Min(min) => to_hash.write_u16(min),
            Constraint::Ratio(left, right) => {
                to_hash.write_u32(left);
                to_hash.write_u32(right);
            }
            Constraint::Length(length) => to_hash.write_u16(length),
            Constraint::Percentage(percentage) => to_hash.write_u16(percentage),
        });
        to_hash.finish()
    }};
}

#[derive(Clone, Copy)]
#[repr(transparent)]
struct CustomHash(u64);

impl Default for CustomHash {
    #[inline]
    fn default() -> Self {
        Self(0)
    }
}

impl std::hash::Hasher for CustomHash {
    #[inline]
    fn finish(&self) -> u64 {
        self.0
    }

    #[inline]
    fn write(&mut self, _: &[u8]) {
        panic!("unsupported operation");
    }

    #[inline]
    fn write_u64(&mut self, i: u64) {
        self.0 = i;
    }
}

thread_local! {
    static LAYOUT_CACHE: RefCell<HashMap<u64, Vec<Rect>, BuildHasherDefault<CustomHash>>> = RefCell::new(HashMap::default());
}

pub fn split(self, area: Rect) -> Vec<Rect> {
    // TODO: Maybe use a fixed size cache ?
    LAYOUT_CACHE.with(|c| {
        c.borrow_mut()
            .entry(hash_layout!(self, area))
            .or_insert_with(|| split(area, self))
            .clone()
    })
}
```
This is the code I used for storing the layout. Here is a breakdown

- First I move the `Layout` into the `split()` function to get more performance out of the hasher
- Then I create a hasher using `ahash`, It's not critical that `ahash` be the hashing algorithm (the default hasher works as well). However, considering how often the `split()` function will be run, I think it's worth the sacrifice of one extra dependency.
- After creating the hasher, I feed it the `area` parameter, and then each field of the `Layout` one by one. `cargo +nightly bench` insisted that this was faster than simply running `self.hash(&mut to_hash);`
- After computing the hash of `area` + `self`, I use the final value as the final key into the `LAYOUT_CAHCE`.
- Because `LAYOUT_CACHE` uses `CustomHash` as it's `RandomState`, the hash computed by `ahash` is simply passed through and no further hashing is done. There's no point in hashing a hash.
- Finally, the resulting `Vec<Rect>` is either grabbed from the `LAYOUT_CACHE` or computed, and then cloned to bypass the reference created by the `HashMap`

## I added some `unsafe` code
If you aren't afraid of a little `unsafe`, then I recommend keeping my `split_ref()` method, this is that method
```rust
pub fn split_ref(self, area: Rect) -> &'static Vec<Rect> {
  LAYOUT_CACHE.with(|c| {
      // SAFETY:
      // "c" is stored in a static variable
      // and therefore has a static lifetime
      unsafe {
          std::mem::transmute::<_, &'static Vec<Rect>>(
              c.borrow_mut()
                  .entry(hash_layout!(self, area))
                  .or_insert_with(|| split(area, self)),
          )
      }
  })
}
```
My reasoning behind the safety of this method is annotated as a comment. This method is *significantly* more performant than the regular `split()` method, as I will demonstrate in the benchmark below

## Benchmarks
I did some benchmarks to see just how much my changes effected performance. This was my benchmark
```rust
#[bench]
fn bench_vertical_split_by_height(b: &mut test::Bencher) {
    let target = Rect {
        x: 2,
        y: 2,
        width: 10,
        height: 10,
    };

    b.iter(|| {
        Layout::default()
            .direction(Direction::Vertical)
            .constraints(
                [
                    Constraint::Percentage(10),
                    Constraint::Max(5),
                    Constraint::Min(1),
                ]
                .as_ref(),
            )
            .split(target)
    });
}
```
When I benchmarked the original implementation and mine, these were my results. (top one is my implementation).
```
test layout::tests::bench_vertical_split_by_height ... bench:          89 ns/iter (+/- 5)
test layout::tests::bench_vertical_split_by_height ... bench:         243 ns/iter (+/- 2)
```
My implementation seems to be ~3x faster than the original implementation.

Swapping out `ahash` with the default hasher (`std::collections::hash_map::DefaultHasher::default()`) gave me this result
```
test layout::tests::bench_vertical_split_by_height ... bench:         148 ns/iter (+/- 10)
```

Using `split_ref()` instead of `split()` got me this result
```
test layout::tests::bench_vertical_split_by_height ... bench:          36 ns/iter (+/- 1)
```
This method is ~2.5x faster than the regular `split()` method, and a whopping ~7x faster than the original implementation

## Smaller changes
I noticed that most of the methods inside of this module (`layout`) were very simple and small, so I made them `const` and `#[inline]`ed them.

## Pre-requisites
All of my changes passed the unit tests and I've added a test to account for the new `split_ref()` method.

Thanks for considering my PR,
      - StratusFearMe21