@StratusFearMe21 2022-01-28
P.S,
Renaming the `split_ref()` function to `split()`, thus removing the original `split()` function entirely, only broke a function in `src/widgets/table.rs`, as well as 2 lines of documentation. If breaking changes aren't important, then it might be a good idea to just remove the original `split()` function and just use the new `split_ref()` function instead