@StratusFearMe21 2022-01-28
Hol, up `split_ref()` causes UB, lemme fix it